+++ src/ewoksorange/bindings/__init__.py
@@ -1,4 +1,3 @@
-from .progress import *  # noqa
 from .owwidgets import *  # noqa
 from .owsconvert import *  # noqa
 from .taskwrapper import *  # noqa